Refactor file and dir names.
