+++ imagenet16/analysis/rsa/rdm_bandpass/compute_activations.py
@@ -6,7 +6,7 @@ import sys
 current_dir = pathlib.Path(os.path.abspath(__file__)).parent
 sys.path.append(str(current_dir) + "/../../../")
 
-from src.analysis.rsa.compute_activations import main
+from src.analysis.rsa.bandpass.compute_activations_single_label import main
 
 if __name__ == "__main__":
     # arguments
