+++ imagenet16/analysis/rsa/rdm_bandpass/compute_bandpass.py
@@ -6,7 +6,7 @@ import sys
 current_dir = pathlib.Path(os.path.abspath(__file__)).parent
 sys.path.append(str(current_dir) + "/../../../")
 
-from src.analysis.rsa.compute_bandpass import main
+from src.analysis.rsa.bandpass.compute_bandpass import main
 
 
 if __name__ == "__main__":
